Test graph + draw on btn press

+++ SteeringCS - Restaurant/SteeringCS/graph/Vertex.cs
@@ -10,10 +10,10 @@ namespace SteeringCS.graph
     {
         public string Name;         //Name of this vertex.
         public List<Edge> Adjacent; //List of adjacent edges.
-        public double Y;            //Y-position on map
-        public double X;            //X-position on map
+        public float Y;            //Y-position on map
+        public float X;            //X-position on map
 
-        public Vertex(string name,double x,double y)
+        public Vertex(string name, float x, float y)
         {
             Name = name;
             X = x;
